backend/requirements.txt
# Core Flask
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
Werkzeug==3.1.3
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1

# Environment
python-dotenv==1.1.1

# OpenAI
openai==1.101.0
anyio==4.10.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
sniffio==1.3.1

# LangChain
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-text-splitters==0.3.9
langsmith==0.4.16
dataclasses-json==0.6.7
jsonpatch==1.33
jsonpointer==3.0.0
marshmallow==3.26.1
mypy_extensions==1.1.0
orjson==3.11.2
tenacity==9.1.2
typing-inspect==0.9.0

# Pydantic (openai, langchain, chromadb)
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.10.1
annotated-types==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.1

# ChromaDB
chromadb==1.0.20
mmh3==5.2.0
overrides==7.7.0
posthog==5.4.0
PyPika==0.48.9
onnxruntime==1.22.1
flatbuffers==25.2.10
msgspec==0.19.0
backoff==2.2.1
rich==14.1.0
typer==0.16.1
PyYAML==6.0.2

# ChromaDB networking
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.4
propcache==0.3.2
yarl==1.20.1

# ChromaDB telemetry
grpcio==1.74.0
grpcio-status==1.71.2
googleapis-common-protos==1.70.0
protobuf==5.29.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0

# Document processing
pdfplumber==0.11.7
pdfminer.six==20250506
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
et_xmlfile==2.0.0
pypdfium2==4.30.0

# Data processing
pandas==2.3.2
numpy==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0

# Networking
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.5.0

# Misc shared deps
cffi==1.17.1
cryptography==45.0.6
pycparser==2.22
filelock==3.19.1
fsspec==2025.7.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.27.0
packaging==25.0
regex==2025.7.34
setuptools==80.9.0
tqdm==4.67.1
zipp==3.23.0
zstandard==0.24.0
coloredlogs==15.0.1
humanfriendly==10.0
filetype==1.2.0
pyparsing==3.2.3